New file transfer code
Have compiled and tested a new file called file buffer, it has
successfully copied both text and jpeg files from one file location to
another using a buffer. my aim for the next stage is to send the buffer
contents down the VLC channel to a second buffer on the client device to
write a new file, effectively copying the file from the main device to
the client device.